Move delete logic out of Gallery and Editor into AssetAdmin
